fix(ci): split source-control-primary-action for max-lines
Extract types, title helpers, and in-flight remote-op resolution into
focused sibling modules so oxlint max-lines (300) passes on the primary
action state machine.
