+++ src/renderer/src/components/right-sidebar/source-control-primary-action.ts
@@ -1,105 +1,28 @@
 // Why: split from the combined primary+dropdown module because the primary and dropdown are independent derivations with different priority ladders; together they exceed the max-lines budget and tangle unrelated concerns.
 
-import type { HostedReviewCreationEligibility } from '../../../../shared/hosted-review'
-import type { GitUpstreamStatus, PRState } from '../../../../shared/types'
 import { shouldForcePushWithLeaseForUpstream } from '../../../../shared/git-upstream-status'
 import { translate } from '@/i18n/i18n'
 import { localizedHostedReviewCopy } from '@/i18n/hosted-review-localized-copy'
+import { type PrimaryAction, type PrimaryActionInputs } from './source-control-primary-action-types'
+import { resolvePrimaryActionDuringRemoteOp } from './source-control-primary-action-in-flight'
+import {
+  describeForcePushWithLease,
+  describePullCount,
+  describePushCount,
+  describeSyncCounts
+} from './source-control-primary-action-titles'
+
+export type {
+  PrimaryActionKind,
+  RemoteOpKind,
+  PrimaryAction,
+  PrimaryActionInputs
+} from './source-control-primary-action-types'
 
 // Why: this module owns the pure state-machine logic for the Source Control
 // primary action (split button). Keeping the logic outside the React component
 // makes it straightforward to unit-test each row of the priority table without
 // spinning up a renderer.
-
-// Why: the primary button collapses to one-label-per-action. Compound
-// kinds ('commit_push', 'commit_sync', 'commit_publish') live in
-// DropdownActionKind only — never on the primary — so they are not part
-// of this union. Narrowing the type here is load-bearing: it lets
-// `handlePrimaryClick` switch exhaustively over only the kinds the
-// primary can actually emit, and it kills the compound-commit branch in
-// the isRemoteOperationActive tooltip below at compile time.
-export type PrimaryActionKind =
-  | 'commit'
-  | 'stage'
-  | 'push'
-  | 'pull'
-  | 'sync'
-  | 'publish'
-  | 'create_pr'
-
-// Why: the in-flight remote op tracker stores which action the user actually
-// triggered, so the primary button can mirror that label/spinner instead of
-// claiming a stale or unrelated operation is running. Dropdown-only remote
-// kinds are included because they participate in the busy flag, but they are
-// intentionally NOT in PrimaryActionKind — when Fetch is in flight the primary
-// keeps its natural label, while Force Push maps back to the push icon/slot.
-export type RemoteOpKind =
-  | 'push'
-  | 'force_push'
-  | 'pull'
-  | 'sync'
-  | 'fetch'
-  | 'fast_forward'
-  | 'publish'
-  | 'rebase'
-
-export type PrimaryAction = {
-  kind: PrimaryActionKind
-  label: string
-  title: string
-  disabled: boolean
-}
-
-export type PrimaryActionInputs = {
-  stagedCount: number
-  hasUnstagedChanges: boolean
-  hasStageableChanges: boolean
-  hasPartiallyStagedChanges: boolean
-  hasMessage: boolean
-  hasUnresolvedConflicts: boolean
-  isCommitting: boolean
-  isRemoteOperationActive: boolean
-  upstreamStatus: GitUpstreamStatus | undefined
-  prState?: PRState | null
-  isPRStateLoading?: boolean
-  // Why: which remote op is currently running, when one is. null when no
-  // remote op is in flight. Used by the in-flight branch below to mirror
-  // the user-triggered action on the primary button instead of leaving a
-  // stale label that no longer matches what the slice is doing.
-  inFlightRemoteOpKind?: RemoteOpKind | null
-  hostedReviewCreation?: HostedReviewCreationEligibility | null
-  // Why: an unpublished branch is only worth publishing when it actually
-  // carries commits beyond the compare base. Undefined preserves the old
-  // behavior while the branch compare request is still unavailable/loading.
-  branchCommitsAhead?: number
-}
-
-const PRIMARY_LABEL_BY_KIND: Record<Exclude<PrimaryActionKind, 'commit'>, string> = {
-  stage: 'Stage All',
-  push: 'Push',
-  pull: 'Pull',
-  sync: 'Sync',
-  publish: 'Publish Branch',
-  create_pr: 'Create PR'
-}
-
-function describePushCount(ahead: number): string {
-  return `Push ${ahead} commit${ahead === 1 ? '' : 's'}`
-}
-
-function describePullCount(behind: number): string {
-  return `Pull ${behind} commit${behind === 1 ? '' : 's'}`
-}
-
-function describeSyncCounts(ahead: number, behind: number): string {
-  return `Pull ${behind}, push ${ahead}`
-}
-
-function describeForcePushWithLease(count: number | undefined, upstreamName?: string): string {
-  const countText =
-    count && count > 0 ? `${count} branch commit${count === 1 ? '' : 's'}` : 'this branch'
-  return `Remote only has older copies of local commits. Force push ${countText} with lease to update ${upstreamName ?? 'the remote branch'}.`
-}
 
 /**
  * Resolve the primary split-button action.
@@ -134,7 +57,6 @@ export function resolvePrimaryAction(inputs: PrimaryActionInputs): PrimaryAction
     upstreamStatus,
     prState,
     isPRStateLoading,
-    inFlightRemoteOpKind,
     hostedReviewCreation,
     branchCommitsAhead
   } = inputs
@@ -155,70 +77,8 @@ export function resolvePrimaryAction(inputs: PrimaryActionInputs): PrimaryAction
     }
   }
 
-  // 2. Remote op in flight — disable the primary. When the in-flight op
-  //    is a primary-eligible kind that doesn't match the primary's natural
-  //    label, mirror the in-flight kind so the user sees the action they
-  //    actually triggered (e.g. "Sync" when they picked Sync from the
-  //    dropdown while the primary's natural state was "Push"). When the
-  //    in-flight op matches the primary's natural kind we keep the natural
-  //    label so its richer detail (counts like "Push 3 commits") survives.
-  //    Fetch and unknown in-flight kinds leave the primary's natural label
-  //    intact; CommitArea's spinner suppresses itself via the kind-mismatch
-  //    check so a non-matching in-flight op doesn't visually claim the
-  //    primary as its host.
   if (isRemoteOperationActive) {
-    const candidate = resolvePrimaryAction({ ...inputs, isRemoteOperationActive: false })
-    const inFlightIsPrimaryKind =
-      inFlightRemoteOpKind === 'push' ||
-      inFlightRemoteOpKind === 'pull' ||
-      inFlightRemoteOpKind === 'sync' ||
-      inFlightRemoteOpKind === 'publish'
-
-    if (inFlightRemoteOpKind === 'force_push') {
-      return {
-        kind: 'push',
-        label: translate(
-          'auto.components.right.sidebar.source.control.primary.action.390abeab93',
-          'Force Push'
-        ),
-        title: translate(
-          'auto.components.right.sidebar.source.control.primary.action.74fc171e99',
-          'Force Push in progress…'
-        ),
-        disabled: true
-      }
-    }
-
-    if (inFlightIsPrimaryKind && candidate.kind !== inFlightRemoteOpKind) {
-      const label = PRIMARY_LABEL_BY_KIND[inFlightRemoteOpKind]
-      return {
-        kind: inFlightRemoteOpKind,
-        label,
-        title: translate(
-          'auto.components.right.sidebar.source.control.primary.action.484f45c439',
-          '{{value0}} in progress…',
-          { value0: label }
-        ),
-        disabled: true
-      }
-    }
-
-    // Why: when the candidate label is "Commit", the generic "remote
-    // operation in progress…" tooltip mismatches the visible label. Point
-    // the user at the fact that the commit will wait, keeping the label and
-    // the explanation consistent. Conflicts take precedence over the remote
-    // tooltip because resolving them is the only action the user can start
-    // while the remote op runs.
-    const title = hasUnresolvedConflicts
-      ? 'Resolve conflicts before committing'
-      : candidate.kind === 'commit'
-        ? 'Remote operation in progress — try again once it finishes'
-        : 'Remote operation in progress…'
-    return {
-      ...candidate,
-      title,
-      disabled: true
-    }
+    return resolvePrimaryActionDuringRemoteOp(inputs, resolvePrimaryAction)
   }
 
   // 3. Unresolved conflicts block any commit path.
